Update HTML template for slide component: move `view-all` div inside ng-container for SldSecOne
The diff shows changes to the HTML template for the slide component in an Angular application. The `view-all` div, which contains a router link to the package list, has been moved inside the `ng-container` for SldSecOne. This means that the `view-all` div will only be rendered when SldSecOne is false. Additionally, some surrounding div elements have been added and removed to accommodate this change.