"Updated CSS and HTML files for blog and slide components; added new method to master service to fetch package data"
This commit includes updates to the CSS and HTML files for the blog and slide components, as well as the addition of a new method to the master service to fetch package data. The specific changes to the CSS and HTML files are detailed in the diff, and the new method in the master service is added to allow for the retrieval of package data in a separate component.